refactor(status): 彻底分离系统状态与任务状态
## 架构变更

**之前**：
- `status.py`: 包含系统状态 + 任务状态混合逻辑
- `vibe3 status` (hidden): 调用 `status.status(show_task_progress=False)`
- `vibe3 task status`: 调用 `status.status()` (默认 True)

**现在**：
- `status.py`: 只包含系统状态逻辑（Orchestra Status + Config）
- `task.py`: 包含完整的任务状态逻辑（Issue Progress + Ready Queue）
- `vibe3 status`: 正式命令，调用 `status.show_system_status()`
- `vibe3 task status`: 独立命令，直接实现任务状态显示

## 代码变更

**status.py** (-242 行):
- 移除 `_include_issue_in_task_progress()` helper
- 移除所有 task progress 渲染逻辑
- 重命名 `status()` → `show_system_status()`
- 添加引导提示

**task.py** (+253 行):
- 实现完整的 task status 逻辑
- 内联 `_include_issue_in_task_progress()` 逻辑
- 直接调用相关服务和渲染函数

**cli.py** (-5 行, +3 行):
- 移除 `hidden=True`，让 `vibe3 status` 成为正式命令
- 移除 `all_flows` 参数（不属于系统状态）
- 调用 `status.show_system_status()`

## 架构优势

1. ✅ **职责分离**：系统状态与任务状态完全解耦
2. ✅ **命令清晰**：`vibe3 status` 与 `vibe3 task status` 并行独立
3. ✅ **易于维护**：各自的逻辑在各自模块中
4. ✅ **符合领域驱动**：系统层 vs 工作层
