+++ src/vibe3/cli.py
@@ -110,9 +110,8 @@ app.add_typer(mcp.app, name="mcp")
 app.add_typer(ask.app, name="ask")
 
 
-@app.command(name="status", hidden=True)
+@app.command(name="status")
 def status_command(
-    all_flows: status.AllOption = False,
     check: Annotated[
         bool,
         typer.Option("--check", help="显示前先运行完整 vibe3 check"),
@@ -129,9 +128,8 @@ def status_command(
         ),
     ] = False,
 ) -> None:
-    """[Compatibility] Redirect to task status."""
-    status.status(
-        all_flows=all_flows,
+    """Show system status: Orchestra and Vibe3 configuration."""
+    status.show_system_status(
         check=check,
         output_format=output_format,
         trace=trace,
